Many changes (see description)
* Commander data will now only be saved if it has actually changed.
* Coordinates for systems have been changed from floats to doubles.
* Epeditions now display "proper" names for certain start types in the scan count list
* Expedition Viewer's elements now properly handle the dialog being resized.
* Expedition Viewer now displays the (game time) timestamp of when a system was added to the expedition log.
* Expedition Viewer can now export visited systems to a text file for importing into programs such as EDDiscovery.
* Expedition Viewer will now highlight system entries for systems that are "landmarks".
* Elite Monitor's main window will now show distances from several in-game landmarks when hovering over the current location text at the top of the window.
* Elite Monitor's main window now provides greater feedback when creating journal entries from unparsed journal files.

+++ EliteMonitor/Elite/Body.cs
@@ -9,11 +9,11 @@ namespace EliteMonitor.Elite
 {
     public class SystemCoordinate
     {
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
 
-        public SystemCoordinate(float X, float Y, float Z)
+        public SystemCoordinate(double X, double Y, double Z)
         {
             this.X = X;
             this.Y = Y;
